feat: enhance global state management by adding pending and accepted profiles, and loading state for friendships
